Modify the default force constant for torsion constraint
Usually a value of 1000 or more is too high for torsional constraint to behave correctly (for the version earlier than 2020.03.3.0). If a torsional scan is run with the constraint, the curve is usually non-smooth. There's a PR in the RDKit seems to fix this issue (for versions after 2021.03), but I had a chance to test it. So I barely lower the value to 100.