Fix single-atom openbabel mol to rdkit mol
Previously, in PR #45, Shih-Cheng mentioned a bug when reading a Gaussian log file with a single atom calculation. It turned out that this was due to issues in converting the perceived OBMol to RDKit Mol. obmol.HasNonZeroCoords() was used as a flag to check if a 3D geometry is included in the OBMol, but it is not helpful for single atom molecule.