Refactor populate_resonance_algorithms
Previously, aromatic agnostic and aromatic specific methods are combined and output. Given the new implementation of PathFinder, it is better to split the two types of methods. The new implementation returns two lists. The first list has the string keys for PathFinders, and the second list has a list of aromatic specific functions