Add a customized get_shortest_path
The RDKit GetShortestPath has a very long setup time ~ 0.5ms (on test machine) regardless of the size of the molecule. As a comparison, on the same machine, a naive python implementation of DFS (`_find_shortest_path`) takes ~0.5 ms for a 100-C normal alkane end to end. Therefore, it make more sense to use a method with a shorter setup time though scaling worse for smaller molecules while using GetShortestPath for larger molecules.