Write input file for gaussian xtb work according to performance suggestion
Here quotes the suggestion: "%NProcShared= will make Gaussian wait busily for xtb using a full half of the NProcShared cores while xtb is running, but use all the cores for it’s own routines (like Hessian diagonalization). It is best to give only 1 or 2 processes to Gaussian and the rest to xtb using -P as above if Hessians are often computed (with xtb parallelization) or the system is large. Otherwise, allocating some threads to Gaussian can make the code faster."